+++ promptview/llms/clients/base.py
@@ -1,0 +1,8 @@
+from typing import Any
+from pydantic import BaseModel
+
+
+class BaseLlmClient(BaseModel):
+    client: Any
+    async def complete(self, msgs, **kwargs):
+        raise NotImplementedError
