+++ promptview/llms/exceptions.py
@@ -1,0 +1,13 @@
+
+
+
+
+
+class LlmError(Exception):
+    pass
+
+
+
+
+class LLMToolNotFound(Exception):
+    pass
